requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.1,<4.0
dj-database-url>=2.1,<3.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
argon2-cffi>=21.3,<22.0
